+++ rxanimation-kotlin/src/main/kotlin/com/reyurnible/rxanimation_kotlin/animation/RxAnimation.kt
@@ -1,9 +1,9 @@
-package com.hosaka.rxanimation_kotlin.animation
+package com.reyurnible.rxanimation_kotlin.animation
 
 import android.view.View
 import android.view.animation.Animation
-import com.hosaka.rxanimation.animation.AnimationEvent
-import com.hosaka.rxanimation.animation.RxAnimation
+import com.reyurnible.rxanimation.animation.AnimationEvent
+import com.reyurnible.rxanimation.animation.RxAnimation
 import rx.Observable
 
 fun Animation.events(view: View): Observable<AnimationEvent> = RxAnimation.events(this, view)
