+++ rxanimation-kotlin/src/main/kotlin/com/reyurnible/rxanimation_kotlin/animation/RxAnimator.kt
@@ -1,11 +1,11 @@
-package com.hosaka.rxanimation_kotlin.animation
+package com.reyurnible.rxanimation_kotlin.animation
 
 import android.animation.Animator
 import android.animation.ValueAnimator
-import com.hosaka.rxanimation.animator.AnimatorEvent
-import com.hosaka.rxanimation.animator.AnimatorPauseEvent
-import com.hosaka.rxanimation.animator.AnimatorUpdateEvent
-import com.hosaka.rxanimation.animator.RxAnimator
+import com.reyurnible.rxanimation.animator.AnimatorEvent
+import com.reyurnible.rxanimation.animator.AnimatorPauseEvent
+import com.reyurnible.rxanimation.animator.AnimatorUpdateEvent
+import com.reyurnible.rxanimation.animator.RxAnimator
 import rx.Observable
 
 fun Animator.events(): Observable<AnimatorEvent> = RxAnimator.events(this)
