+++ rxanimation/src/androidTest/java/com/reyurnible/rxanimation/RecordingObserver.java
@@ -1,4 +1,4 @@
-package com.hosaka.rxanimation;
+package com.reyurnible.rxanimation;
 
 import android.util.Log;
