+++ rxanimation/src/main/java/com/reyurnible/rxanimation/animator/AnimatorPauseEvent.java
@@ -1,28 +1,27 @@
-package com.hosaka.rxanimation.animator;
+package com.reyurnible.rxanimation.animator;
 
 import android.animation.Animator;
 import android.support.annotation.CheckResult;
 import android.support.annotation.NonNull;
 
 /**
- * A android animator event occurred.
+ * A android animator pause event occurred.
  */
-public class AnimatorEvent {
+public class AnimatorPauseEvent {
 
     public enum Kind {
-        START, END, CANCEL, REPEAT
+        PAUSE, RESUME
     }
 
-    @CheckResult
-    @NonNull
-    public static AnimatorEvent create(@NonNull Animator animation, @NonNull Kind kind) {
-        return new AnimatorEvent(animation, kind);
+    @CheckResult @NonNull
+    public static AnimatorPauseEvent create(@NonNull Animator animation, @NonNull Kind kind) {
+        return new AnimatorPauseEvent(animation, kind);
     }
 
     private final Kind kind;
     private Animator animation;
 
-    private AnimatorEvent(Animator animation, Kind kind) {
+    private AnimatorPauseEvent(Animator animation, Kind kind) {
         this.animation = animation;
         this.kind = kind;
     }
@@ -37,26 +36,23 @@ public class AnimatorEvent {
         return kind;
     }
 
-    @Override
-    public boolean equals(Object object) {
+    @Override public boolean equals(Object object) {
         if (object == this) return true;
-        if (!(object instanceof AnimatorEvent)) return false;
-        AnimatorEvent other = (AnimatorEvent) object;
+        if (!(object instanceof AnimatorPauseEvent)) return false;
+        AnimatorPauseEvent other = (AnimatorPauseEvent) object;
         return other.animation() == animation()
                 && other.kind() == kind();
     }
 
-    @Override
-    public int hashCode() {
+    @Override public int hashCode() {
         int result = 17;
         result = result * 37 + animation().hashCode();
         result = result * 37 + kind().hashCode();
         return result;
     }
 
-    @Override
-    public String toString() {
-        return "AnimatorEvent{animation="
+    @Override public String toString() {
+        return "AnimatorPauseEvent{animation="
                 + animation()
                 + ", kind="
                 + kind()
